requirements.txt
matplotlib==3.6.2
numpy==1.24.4
scikit_learn==1.1.3
setuptools==65.5.0
sympy==1.11.1
paddlepaddle-gpu==2.2.2
tqdm==4.66.2
